@@ -2,20 +2,27 @@ package com.lukelorusso.simplepaperviewsample
 
 import android.os.Bundle
 import android.os.Handler
+import android.os.Looper
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.ContextCompat
 import androidx.core.content.res.ResourcesCompat
 import com.lukelorusso.simplepaperview.SimplePaperView
-import kotlinx.android.synthetic.main.activity_main.*
+import com.lukelorusso.simplepaperviewsample.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
+    private lateinit var binding: ActivityMainBinding
     private lateinit var colors: List<Int>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        ActivityMainBinding.inflate(layoutInflater).also { inflated ->
+            binding = inflated
+            setContentView(binding.root)
+        }
 
         //region INIT VIEW
         colors = listOf(
@@ -26,10 +33,14 @@ class MainActivity : AppCompatActivity() {
         )
 
         drawStuff()
-        mainSimplePaperView.setOnDrawListener { mainScrollView.post { mainScrollView.fullScroll(View.FOCUS_RIGHT) } }
-        mainBtnClear.setOnClickListener { mainSimplePaperView.clearPaper() }
-        mainBtnRedraw.setOnClickListener {
-            mainSimplePaperView.clearPaper(false)
+        binding.mainSimplePaperView.setOnDrawListener {
+            binding.mainScrollView.post {
+                binding.mainScrollView.fullScroll(View.FOCUS_RIGHT)
+            }
+        }
+        binding.mainBtnClear.setOnClickListener { binding.mainSimplePaperView.clearPaper() }
+        binding.mainBtnRedraw.setOnClickListener {
+            binding.mainSimplePaperView.clearPaper(false)
             drawStuff()
         }
         //endregion
@@ -37,14 +48,14 @@ class MainActivity : AppCompatActivity() {
 
     private fun drawStuff() {
         drawText()
-        Handler().postDelayed({ draw1stItem() }, 500)
-        Handler().postDelayed({ draw2ndItem() }, 1000)
-        Handler().postDelayed({ draw3thItem() }, 1500)
-        Handler().postDelayed({ draw4thItem() }, 2000)
+        Handler(Looper.getMainLooper()).postDelayed({ draw1stItem() }, 500)
+        Handler(Looper.getMainLooper()).postDelayed({ draw2ndItem() }, 1000)
+        Handler(Looper.getMainLooper()).postDelayed({ draw3thItem() }, 1500)
+        Handler(Looper.getMainLooper()).postDelayed({ draw4thItem() }, 2000)
     }
 
     private fun drawText() {
-        mainSimplePaperView.drawInDp(
+        binding.mainSimplePaperView.drawInDp(
             SimplePaperView.TextLabel(
                 getString(R.string.textLabel),
                 18F,
@@ -58,28 +69,33 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun draw1stItem() {
-        mainSimplePaperView.drawInDp(
+        binding.mainSimplePaperView.drawInDp(
             SimplePaperView.Circle(110F, 100F, 25F, colors[0])
         )
     }
 
     private fun draw2ndItem() {
-        mainSimplePaperView.drawInDp(
+        binding.mainSimplePaperView.drawInDp(
             SimplePaperView.Line(200F, 200F, 400F, 0F, colors[1], 8F)
         )
     }
 
     private fun draw3thItem() {
-        mainSimplePaperView.drawInDp(
+        binding.mainSimplePaperView.drawInDp(
             SimplePaperView.Line(410F, 60F, 580F, 130F, colors[2], 2F)
         )
     }
 
     private fun draw4thItem() {
-        mainSimplePaperView.drawInDp(
+        binding.mainSimplePaperView.drawInDp(
             listOf(
                 SimplePaperView.Circle(660F, 130F, 25F, colors[3]),
-                SimplePaperView.Circle(660F, 130F, 23F, mainSimplePaperView.getBackgroundColor())
+                SimplePaperView.Circle(
+                    660F,
+                    130F,
+                    23F,
+                    binding.mainSimplePaperView.getBackgroundColor()
+                )
             )
         )
     }
